Refactors code block tab indentation to a hook
