+++ packages/block-library/src/code/edit.js
@@ -3,8 +3,12 @@
  */
 import { RichText, useBlockProps } from '@wordpress/block-editor';
 import { createBlock, getDefaultBlockName } from '@wordpress/blocks';
-import { useRefEffect } from '@wordpress/compose';
 import { __ } from '@wordpress/i18n';
+
+/**
+ * Internal dependencies
+ */
+import { useTabIndentation } from './hook';
 
 export default function CodeEdit( {
 	attributes,
@@ -14,75 +18,7 @@ export default function CodeEdit( {
 	mergeBlocks,
 } ) {
 	const blockProps = useBlockProps();
-
-	const ref = useRefEffect( ( element ) => {
-		function handleKeyDown( event ) {
-			if ( event.key === 'Tab' ) {
-				const ownerDocument = element.ownerDocument;
-				const selection = ownerDocument.defaultView.getSelection();
-
-				if ( selection.rangeCount > 0 ) {
-					const range = selection.getRangeAt( 0 );
-
-					if ( ! range.collapsed ) {
-						event.preventDefault();
-						event.stopPropagation();
-
-						const selectedText = getTextWithLineBreaks( range );
-						let modifiedText = selectedText;
-
-						if ( event.shiftKey ) {
-							// Shift+Tab: Remove tab indentation
-							modifiedText = selectedText
-								.split( '\n' )
-								.map( ( line ) => {
-									if ( line.startsWith( '\t' ) ) {
-										return line.substring( 1 );
-									}
-									return line;
-								} )
-								.join( '\n' );
-						} else {
-							// Tab: Add indentation
-							modifiedText = selectedText
-								.split( '\n' )
-								.map( ( line ) =>
-									line.length > 0 ? '\t' + line : line
-								)
-								.join( '\n' );
-						}
-
-						range.deleteContents();
-						const textNode =
-							ownerDocument.createTextNode( modifiedText );
-						range.insertNode( textNode );
-
-						const newRange = ownerDocument.createRange();
-						newRange.selectNode( textNode );
-						selection.removeAllRanges();
-						selection.addRange( newRange );
-					}
-				}
-			}
-		}
-
-		function getTextWithLineBreaks( range ) {
-			const fragment = range.cloneContents();
-			const tempDiv = document.createElement( 'div' );
-			tempDiv.appendChild( fragment );
-
-			tempDiv.querySelectorAll( 'br, div, p, li' ).forEach( ( el ) => {
-				el.insertAdjacentText( 'beforebegin', '\n' );
-			} );
-
-			return tempDiv.textContent.replace( /\n$/, '' );
-		}
-
-		element.addEventListener( 'keydown', handleKeyDown, true );
-		return () => {
-			element.removeEventListener( 'keydown', handleKeyDown, true );
-		};
-	}, [] );
+	const { ref } = useTabIndentation();
 
 	return (
 		<pre { ...blockProps }>
